+++ templates/sidebar-proud-topic.php
@@ -1,0 +1,1 @@
+<?php dynamic_sidebar('sidebar-proud-topic'); ?><!-- template-file: templates/sidebar-proud-topic.php -->